requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
redis==5.0.8
sqlalchemy==2.0.35
apscheduler==3.10.4
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
